Move Finish component to containers directory
